implements supplier address and product relation removal on bulk delete action

+++ src/Adapter/Supplier/CommandHandler/BulkDeleteSupplierHandler.php
@@ -28,7 +28,9 @@ namespace PrestaShop\PrestaShop\Adapter\Supplier\CommandHandler;
 
 use PrestaShop\PrestaShop\Core\Domain\Supplier\Command\BulkDeleteSupplierCommand;
 use PrestaShop\PrestaShop\Core\Domain\Supplier\CommandHandler\BulkDeleteSupplierHandlerInterface;
+use PrestaShop\PrestaShop\Core\Domain\Supplier\Exception\CannotDeleteSupplierAddressException;
 use PrestaShop\PrestaShop\Core\Domain\Supplier\Exception\CannotDeleteSupplierException;
+use PrestaShop\PrestaShop\Core\Domain\Supplier\Exception\CannotDeleteSupplierProductRelationException;
 use PrestaShop\PrestaShop\Core\Domain\Supplier\Exception\SupplierException;
 use PrestaShop\PrestaShop\Core\Domain\Supplier\Exception\SupplierNotFoundException;
 use PrestaShopException;
@@ -37,7 +39,7 @@ use Supplier;
 /**
  * Class BulkDeleteSupplierHandler is responsible for deleting multiple suppliers.
  */
-final class BulkDeleteSupplierHandler implements BulkDeleteSupplierHandlerInterface
+final class BulkDeleteSupplierHandler extends AbstractDeleteSupplierHandler implements BulkDeleteSupplierHandlerInterface
 {
     /**
      * {@inheritdoc}
@@ -59,7 +61,44 @@ final class BulkDeleteSupplierHandler implements BulkDeleteSupplierHandlerInterf
                     );
                 }
 
+                if ($this->hasPendingOrders($supplierId)) {
+                    throw new CannotDeleteSupplierException(
+                        $supplierId,
+                        sprintf(
+                            'Supplier with id %s cannot be deleted due to it has pending orders',
+                            $supplierId->getValue()
+                        ),
+                        CannotDeleteSupplierException::HAS_PENDING_ORDERS
+                    );
+                }
+
+                $this->startTransaction();
+
+                if (false === $this->deleteProductSupplierRelation($supplierId)) {
+                    $this->rollbackTransaction();
+
+                    throw new CannotDeleteSupplierProductRelationException(
+                        sprintf(
+                            'Unable to delete suppliers with id "%s" product relation from product_supplier table',
+                            $supplierId->getValue()
+                        )
+                    );
+                }
+
+                if (false === $this->deleteSupplierAddress($supplierId)) {
+                    $this->rollbackTransaction();
+
+                    throw new CannotDeleteSupplierAddressException(
+                        sprintf(
+                            'Unable to set deleted flag for supplier with id "%s" address',
+                            $supplierId->getValue()
+                        )
+                    );
+                }
+
                 if (false === $entity->delete()) {
+                    $this->rollbackTransaction();
+
                     throw new CannotDeleteSupplierException(
                         $supplierId,
                         sprintf(
@@ -68,6 +107,8 @@ final class BulkDeleteSupplierHandler implements BulkDeleteSupplierHandlerInterf
                         )
                     );
                 }
+
+                $this->commitTransaction();
             }
         } catch (PrestaShopException $exception) {
             throw new SupplierException(
